+++ js/utils/constants.js
@@ -1,0 +1,4 @@
+// Multi-file edit test
+export const MAX_RETRIES = 3;
+export const RETRY_DELAY_MS = 2000;
+export const REQUEST_TIMEOUT_MS = 180000;